changed technology ordering
ordering of data.json is now based on (ring, name). This is due to the rendering of the techradar. In the expanded view for each quadrant, the blips are added alphabetically. This leads to seperate orderings of the rings for each quadrant. Which is confusing for end users. Therefore, sort the data.json file in the export function of the Excel file.